requirements.txt
# AWS and Core Dependencies
boto3==1.28.36
python-dotenv==1.0.0
streamlit==1.26.0

# API Requests and HTTP
requests==2.31.0
aiohttp==3.8.5

# Image Processing
Pillow==10.0.0

# Async Support
asyncio==3.4.3

# Instagram API
facebook-sdk==3.1.0

# Utility Packages
python-dateutil==2.8.2
tqdm==4.66.1
